Move ArrowTestDataUtil out of exported package
Move ArrowTestDataUtil out of a module-exported package so that it can
be reused properly in other tests.
